refactor(reticulum): move LXMF send helpers out of useSendMessage
Keep the hook as a thin toast/adapter while destination resolution, reply
fields, and outbound persistence live under lib/reticulum.
